// Source/OrderKing.Management/src/app/model/user/userinfo.model.ts
export class UserInfoModel {
    UserId: string;
    StoreName: '';
    AccountName: '';
    Password: String;
    FullName: '';
    Email: '';
    PhoneNumber: '';
    Address: '';
    Address2: '';
    IdentityCard: '';
    JoinDate: string;
    CreatedDate: Date;
    IsActived: boolean;
    Avatar: string;

    constructor() {
      this.IsActived = true;
    }
}